处理日K线 昨日收盘价、涨跌幅、换手率 数据

+++ trade-model-tradecore/src/main/java/com/trade/model/tradecore/kline/KLineBase.java
@@ -46,7 +46,7 @@ public class KLineBase implements Serializable {
 	/**
 	 * 成交额
 	 */
-	private float turnover;
+	private long turnover;
 
 	/**
 	 * 换手率
@@ -139,11 +139,11 @@ public class KLineBase implements Serializable {
 		this.volume = volume;
 	}
 
-	public float getTurnover() {
+	public long getTurnover() {
 		return turnover;
 	}
 
-	public void setTurnover(float turnover) {
+	public void setTurnover(long turnover) {
 		this.turnover = turnover;
 	}
